# Conflicts:
#	src/main/java/com/example/pokeplushback/Controladores/ProductosController.java

+++ src/main/java/com/example/pokeplushback/Dto/ProductosDTO.java
@@ -2,6 +2,8 @@ package com.example.pokeplushback.Dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -10,7 +12,7 @@ public class ProductosDTO {
     private Integer id;
     private String nombre;
     private String descripcion;
-    private float precio;
+    private BigDecimal precio;
     private String tipo;
     private Long foto;
     private int stock;
